src/linker.cpp: Fix conflicting `path` definition
This applies the patch from but report [9] to fix the conflicting
`path` global between m100emu.cpp and linker.cpp.

This is a fairly old bug; the patch was submitted 2022-11-08. It's not
clear why this hasn't been applied, but the code seems to be seeing
minimal maintenance. (There's been only one commit since 2021-05-22.

[9]: https://sourceforge.net/p/virtualt/bugs/9/